requirements.txt
streamlit==1.46.0
streamlit-lottie==0.0.5
streamlit-option-menu==0.4.0
pandas==2.3.0
numpy==1.26.4
scikit-learn==1.6.1
scipy==1.15.3
matplotlib==3.10.3
plotly==6.1.2
pyod==2.0.5
tensorflow==2.16.1
keras==3.10.0
# Utilidades extra
joblib==1.5.1
requests==2.32.4
psutil==7.0.0
combo==0.1.3

# Reportes PDF
fpdf2>=2.7.0
